+++ src/providers/LocalAudioOutputProvider/index.tsx
@@ -12,10 +12,12 @@ import React, {
 } from 'react';
 import { LocalAudioOutputContextType } from '../../types';
 import { useAudioVideo } from '../AudioVideoProvider';
+import { useLogger } from '../LoggerProvider';
 
 const Context = createContext<LocalAudioOutputContextType | null>(null);
 
 const LocalAudioOutputProvider: React.FC = ({ children }) => {
+  const logger = useLogger();
   const audioVideo = useAudioVideo();
   const [isAudioOn, setIsAudioOn] = useState(true);
   const audioRef = useRef<HTMLAudioElement>(null);
@@ -30,7 +32,7 @@ const LocalAudioOutputProvider: React.FC = ({ children }) => {
         try {
           await audioVideo.bindAudioElement(element);
         } catch (e) {
-          console.error('Failed to bind audio element.', e);
+          logger?.error(`Failed to bind audio element. ${e}`);
         }
       })(audioRef.current);
     }
@@ -52,7 +54,7 @@ const LocalAudioOutputProvider: React.FC = ({ children }) => {
         try {
           await audioVideo?.bindAudioElement(element);
         } catch (e) {
-          console.error('Failed to bind audio element.', e);
+          logger?.error(`Failed to bind audio element. ${e}`);
         }
       })(audioRef.current);
     }